+++ spring-cloud-gcp-autoconfigure/pom.xml
@@ -219,19 +219,20 @@
             <artifactId>spring-security-config</artifactId>
             <optional>true</optional>
         </dependency>
-		
+
         <dependency>
             <groupId>org.springframework.security</groupId>
             <artifactId>spring-security-oauth2-resource-server</artifactId>
             <optional>true</optional>
         </dependency>
 				
-		<!-- Actuator -->
-		<dependency>
-	  		<groupId>org.springframework.boot</groupId>
-	  		<artifactId>spring-boot-starter-actuator</artifactId>
-		</dependency>
-        
+        <!-- Actuator -->
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-actuator</artifactId>
+            <optional>true</optional>
+        </dependency>
+
         <!-- Test -->
         <dependency>
             <groupId>commons-io</groupId>
