--- spring-cloud-gcp-autoconfigure/src/test/java/org/springframework/cloud/gcp/autoconfigure/storage/Content.java
@@ -1,41 +1,0 @@
-/*
- * Copyright 2018-2019 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * Test class for storage.
- *
- * @author Raghavan N S
- * @author Srinivasa Meenavalli
- */
-
-package org.springframework.cloud.gcp.autoconfigure.storage;
-
-import java.io.Serializable;
-
-public class Content implements Serializable {
-
-	private static final long serialVersionUID = 1L;
-	private Long contentLength;
-
-	public Long getContentLength() {
-		return contentLength;
-	}
-
-	public void setContentLength(Long contentLength) {
-		this.contentLength = contentLength;
-	}
-
-}
