+++ spring-cloud-gcp-data-datastore/pom.xml
@@ -41,6 +41,6 @@
     <dependency>
       <groupId>org.springframework</groupId>
       <artifactId>spring-tx</artifactId>
-    </dependency>	
+    </dependency>
   </dependencies>
 </project>
